# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: NS Scrap ERP Next API
  version: 0.1.0
  description: >
    Baseline API catalog for the active Next.js app. This file records current
    route handlers first; detailed request and response schemas should be added
    as each module batch hardens its contract.
servers:
  - url: http://localhost:3000
    description: Local Next.js development server
  - url: https://new-ns-scrap-erp.vercel.app
    description: Current Vercel deployment, if available
security:
  - SupabaseSession: []
...
paths:
...
  /api/stock/convert:
    get:
      tags: [Stock]
      operationId: listStockConverts
      summary: List grade/product conversion history
      responses:
        "200":
          description: Product conversion rows and stock reference options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockOperationListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Stock]
      operationId: createStockConvert
      summary: Convert stock from one product/grade to another
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockConvertRequest"
      responses:
        "200":
          $ref: "#/components/responses/StockWriteOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/stock/adjust:
    get:
      tags: [Stock]
      operationId: listStockAdjustments
      summary: List stock count adjustments
      responses:
        "200":
          description: Stock count adjustment rows and stock reference options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockOperationListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Stock]
      operationId: createStockAdjustment
      summary: Create stock count adjustment and ledger movement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StockAdjustRequest"
      responses:
        "200":
          $ref: "#/components/responses/StockWriteOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/stock/customer-return:
    get:
      tags: [Stock]
      operationId: listCustomerReturnStock
      summary: List customer return stock
      responses:
        "200":
          description: Customer return stock rows and stock reference options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StockOperationListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Stock]
      operationId: createCustomerReturnStockMovement
      summary: Receive or send back customer return stock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerReturnRequest"
      responses:
        "200":
          $ref: "#/components/responses/StockWriteOk"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    SupabaseSession:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase Auth session cookie or equivalent authenticated request context.
...
  responses:
...
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid Supabase session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    StockWriteOk:
      description: Stock write completed and traceable ledger rows were created
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/StockWriteResponse"
  schemas:
...
    CustomerReturnRequest:
      allOf:
        - $ref: "#/components/schemas/StockBaseWriteRequest"
        - type: object
          required: [action, branchId, date, productId, qty, reason, unitCost, warehouseId]
          properties:
            action:
              type: string
              enum: [receive, send_back]
              default: receive
            customerId:
              type: [string, "null"]
            deliveryRefNo:
              $ref: "#/components/schemas/OptionalDocNo"
            qty:
              type: number
              exclusiveMinimum: 0
            reason:
              type: string
              minLength: 2
              maxLength: 240
            returnRowKey:
              type: [string, "null"]
            unitCost:
              type: number
              minimum: 0
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      additionalProperties: true
    OptionalDocNo:
      type: [string, "null"]
      maxLength: 40
      pattern: "^[A-Za-z0-9_-]+$"
    StockAdjustRequest:
      allOf:
        - $ref: "#/components/schemas/StockBaseWriteRequest"
        - type: object
          required: [branchId, countedQty, date, productId, reason, systemQty, warehouseId]
          properties:
            countedQty:
              type: number
              minimum: 0
            reason:
              type: string
              minLength: 3
              maxLength: 240
            remark:
              type: [string, "null"]
              maxLength: 500
            systemQty:
              type: number
              minimum: 0
...
    StockBaseWriteRequest:
      type: object
      required: [branchId, date, productId, warehouseId]
      properties:
        branchId:
          type: string
          minLength: 1
        date:
          type: string
          format: date
        docNo:
          $ref: "#/components/schemas/OptionalDocNo"
        lotNo:
          type: [string, "null"]
          maxLength: 80
        notes:
          type: [string, "null"]
          maxLength: 500
        productId:
          type: string
          minLength: 1
        warehouseId:
          type: string
          minLength: 1
    StockConvertRequest:
      type: object
      required: [branchId, date, sourceProductId, sourceQty, targetProductId, targetQty, warehouseId]
      properties:
        branchId:
          type: string
          minLength: 1
        date:
          type: string
          format: date
        docNo:
          $ref: "#/components/schemas/OptionalDocNo"
        lotNo:
          type: [string, "null"]
          maxLength: 80
        notes:
          type: [string, "null"]
          maxLength: 500
        reason:
          type: [string, "null"]
          maxLength: 240
        sourceProductId:
          type: string
          minLength: 1
        sourceQty:
          type: number
          exclusiveMinimum: 0
        targetLotNo:
          type: [string, "null"]
          maxLength: 80
        targetProductId:
          type: string
          minLength: 1
        targetQty:
          type: number
          exclusiveMinimum: 0
        warehouseId:
          type: string
          minLength: 1
...
    StockOption:
      type: object
      required: [id, name]
      properties:
        active:
          type: [boolean, "null"]
        branchId:
          type: [string, "null"]
        code:
          type: [string, "null"]
        id:
          type: string
        name:
          type: string
        status:
          type: [string, "null"]
    StockOperationListResponse:
      type: object
      required: [reference, rows]
      properties:
        reference:
          type: object
          required: [branches, products, warehouses]
          properties:
            branches:
              type: array
              items:
                $ref: "#/components/schemas/StockOption"
            customers:
              type: array
              items:
                $ref: "#/components/schemas/StockOption"
            products:
              type: array
              items:
                $ref: "#/components/schemas/StockOption"
            warehouses:
              type: array
              items:
                $ref: "#/components/schemas/StockOption"
        rows:
          type: array
          items:
            type: object
            additionalProperties: true
...
    StockWriteResponse:
      type: object
      required: [ok]
      properties:
        id:
          type: string
        ok:
          type: boolean
        refId:
          type: string
        refNo:
          type: string
...